Stock_market_projappl/requirements.txt
streamlit==1.25.0
pandas==1.5.3
mysql-connector-python==8.0.33
gspread==5.7.2
google-auth==2.22.0
streamlit-authenticator==0.4.3
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
